Fixed a couple of compiler warnings.

+++ external/pagedgeometry/source/BatchPage.cpp
@@ -115,7 +115,7 @@ void BatchPage::build()
 		//the same as the shader (for computers with no shader support)
 		for (size_t t = 0; t < mat->getNumTechniques(); ++t){
 			Technique *tech = mat->getTechnique(t);
-			for (int p = 0; p < tech->getNumPasses(); ++p){
+			for (size_t p = 0; p < tech->getNumPasses(); ++p){
 				Pass *pass = tech->getPass(p);
 				if (pass->getVertexProgramName().empty())
 					pass->setSpecular(0, 0, 0, 1);
